paas2/appengine/requirements.txt
Django==1.8.19
djangorestframework==3.3.3
Markdown==2.6.6
requests==2.21.0
PyYAML==5.1
uWSGI==2.0.13.1
pymysql==0.6.7
gunicorn==19.9.0
gevent==1.1.2
pytz==2016.6.1
six==1.12.0
python-consul==1.1.0

# for config
django-environ==0.4.5
